Crate imxrt_ccm

Clock Control Module (CCM) driver for i.MX RT systems

imxrt-ccm let you configure clocks, associate peripherals with clock gates, and control peripheral clock gates. It is a lower-level driver, targeted for HAL implementers. We encourage you to re-export imxrt-ccm APIs in your larger libraries.

The rest of this documentation is for HAL implementers, or users who want to create higher-level abstractions which require CCM configuration. HAL users who are using re-exported imxrt-ccm APIs should consult their HAL documentation for more information on clock configuration.

Usage

Implement Instance for your peripheral instances. It's your decision on what qualifies as a "peripheral instance." It could be a type that could represent the register block. Or, it could represent your peripheral driver.

This CCM driver supports clock gating for a variety of peripherals; look for "peripheral instance" in documentation. You should encapsulate any chip-specific details in the Instance implementation.

Then, create a safe wrapper around CCM::new so that your users can safely acquire the CCM peripheral. Your API must ensure that there is only one CCM instance. The types wrapped by your CCM clocks should reflect your Instance implementations.

Here's an example of how to implement an I2C Instance for compatibility with the I2C clock. The example shows how you might include support for the two extra I2C peripherals that are available on a 1060 chip family.

use imxrt_ccm as ccm;

/// Our I2C instance
struct I2C {
    instance_id: usize,
    // Other members...
}

unsafe impl ccm::Instance for I2C {
    type Inst = ccm::i2c::I2C;
    fn instance(&self) -> Self::Inst {
        match self.instance_id {
            1 => ccm::i2c::I2C::I2C1,
            2 => ccm::i2c::I2C::I2C2,
            #[cfg(feature = "imxrt1060")]
            3 => ccm::i2c::I2C::I2C3,
            #[cfg(feature = "imxrt1060")]
            4 => ccm::i2c::I2C::I2C4,
            _ => unreachable!()
        }
    }
    fn is_valid(inst: Self::Inst) -> bool {
        #[allow(unreachable_patterns)]
        match inst {
            ccm::i2c::I2C::I2C1 | ccm::i2c::I2C::I2C2 => true,
            #[cfg(feature = "imxrt1060")]
            ccm::i2c::I2C::I2C3 | ccm::i2c::I2C::I2C4 => true,
            _ => false,
        }
    }
}

type CCM = ccm::CCM<
    // Other clock types...
    I2C
>;

fn take_ccm() -> Option<CCM> {
  // TODO safety check that ensures
  // CCM only taken once!
  Some(unsafe { CCM::new() })
}

let CCM{ mut handle, i2c_clock, .. } = take_ccm().unwrap();
// Enable the clock, which disables all clock gates
let mut i2c_clock = i2c_clock.enable(&mut handle);

We recommend that you create driver initialization APIs that require clocks. By requiring an immutable clock, you guarantee that a user has enabled the peripheral clock in their code.

struct I2CDriver {
    inst: I2C,
    // ...
}

impl I2CDriver {
    pub fn new(inst: I2C, clock: &ccm::i2c::I2CClock<I2C>) -> I2CDriver {
        // ...
        I2CDriver {
            inst,
            // ...
        }
    }
}

let mut i2c3 = // Get I2C3 instance...
// Enable I2C3 clock gate
i2c_clock.set_clock_gate(&mut i2c3, ccm::ClockGate::On);
// Create the higher-level driver, requires the I2C clock
let i2c = I2CDriver::new(i2c3, &i2c_clock);

imxrt-ral support

imxrt-ccm provides support for imxrt-ral. The feature includes Instance implementations on all supported imxrt-ral peripheral instances. It also include helper functions and types, which are exported in the ral module. Use the imxrt-ral support if your HAL already depends on the imxrt-ral crate.

To use the imxrt-ral support, enable the imxrt-ral feature. You must ensure that something else in your dependency graph enables the correct imxrt-ral feature for your processor. See the imxrt-ral documentation for more information.

Chip support

imxrt-ccm does not require you to select a chip. If you do not select a chip, the crate provides the most conservative implementation to support all i.MX RT vairants. However, imxrt-ccm has i.MX RT chip features to specialize the CCM driver for your system. You should enable one of these features in your final program, but it's not required.

The table below describes imxrt-ccm chip support.

Feature	Description

"imxrt1010"	Support for i.MX RT 1010 processors, like iMXRT1011

"imxrt1060"	Support for i.MX RT 1060 processors, like iMXRT1061 and iMXRT1062

If you enable the imxrt-ral feature, you must enable one of these features.

Modules

arm
ARM clock control

i2c
I2C clock control

perclock
Periodic clock

ral	imxrt-ral
Implementations for the imxrt-ral

spi
SPI clock control

uart
UART clock control

Structs

CCM
The root clocks and CCM handle

ClockGateLocation
Describes the location of a clock gate field

DCDC
Peripheral instance identifier for DCDC

DMA
Peripheral instance identifier for DMA

Disabled
A disabled clock of type Clock

Handle
Handle to the CCM register block

Enums

ADC
Peripheral instance identifier for ADCs

ClockGate
Describes a clock gate setting

PWM
Peripheral instance identifier for PWM

Traits

ClockGateLocator
A type that can locate a clock gate

Instance
A peripheral instance that has a clock gate

Functions

get_clock_gate
Returns the clock gate setting for a peripheral instance

set_clock_gate⚠
Set the clock gate for a peripheral instance