Struct imxrt_ccm::perclock::PerClock

pub struct PerClock<P, G>(_);

The periodic clock root

PerClock is the input clock for GPT and PIT.

Implementations

impl<P, G> PerClock<P, G>

pub fn frequency(&self) -> u32

Returns the configured periodic clock frequency

The method requires a reference to the CCM Handle, since it may need to read the IPG clock frequency.

pub fn try_frequency(&self) -> Option<u32>

Try to read the periodic clock frequency, returning the frequency if it can be safely read

If the periodic clocks run on the IPG frequency, it is not safe to read the frequencies. try_frequency would return None. But, if the periodic clocks run on the oscillator, we can safely compute the frequency.

pub fn selection(&self) -> Selection

Returns the periodic clock selection

impl<P, G> PerClock<P, G> where
G: Instance<Inst = GPT>,

pub fn clock_gate_gpt(&self, gpt: &G) -> ClockGate

Returns the clock gate setting for the GPT

pub fn set_clock_gate_gpt(&mut self, gpt: &mut G, gate: ClockGate)

Set the clock gate for the GPT

impl<P, G> PerClock<P, G> where
P: Instance<Inst = PIT>,

pub fn clock_gate_pit(&self, pit: &P) -> ClockGate

Returns the clock gate setting for the PIT

pub fn set_clock_gate_pit(&mut self, pit: &mut P, gate: ClockGate)

Set the clock gate for the PIT

impl<P, G> PerClock<P, G> where
P: Instance<Inst = PIT>,
G: Instance<Inst = GPT>,

pub fn configure_selection_divider(
&mut self,
selection: Selection,
divider: u32
)

Configure the periodic clock root, specifying the clock divider

The divider should be between [1, 64]. The function will treat a 0 as 1, and anything greater than 64 as 64.

When configure returns, all GPT and PIT clock gates will be set to off. To re-configure clock gates, use the clock gate methods on PerClock.

pub fn configure(&mut self)

Configure the periodic clock root with a default divider. The default divider will result in a periodic clock frequency of 1MHz from the crystal oscillator.

When configure returns, all GPT and PIT clock gates will be set to off. To re-configure clock gates, use the clock gate methods on PerClock.