Struct imxrt_ccm::CCM

#[non_exhaustive]pub struct CCM<C: Clocks> { /* fields omitted */ }

The clock control module (CCM)

Implementations

impl CCM<Clocks>

pub fn from_ral(_: Instance) -> Self

This is supported on crate feature imxrt-ral only.

Converts the imxrt-ral CCM instance into the CCM driver

This is safer than using new(), since we take ownership of the only other CCM instance in the system.

use imxrt_ccm::CCM;
use imxrt_ral::ccm;

let ccm = ccm::CCM::take().map(CCM::from_ral).unwrap();

impl<C: Clocks> CCM<C>

pub unsafe fn new() -> Self

Construct a new CCM peripheral

Safety

This should only be called once. Ideally, it's encapsulated behind another constructor that takes ownership of CCM peripheral memory. Calling this more than once will let you access global, mutable memory that's assumed to not be aliased.

pub fn clock_gate_dcdc<D>(&self, dcdc: &D) -> ClockGate where
D: Instance<Inst = DCDC>,

Returns the clock gate setting for the DCDC buck converter

pub fn set_clock_gate_dcdc<D>(&mut self, dcdc: &mut D, gate: ClockGate) where
D: Instance<Inst = DCDC>,

Set the clock gate for the DCDC buck converter

pub fn clock_gate_dma<D>(&self, dma: &D) -> ClockGate where
D: Instance<Inst = DMA>,

Returns the clock gate setting for the DMA controller

pub fn set_clock_gate_dma<D>(&mut self, dma: &mut D, gate: ClockGate) where
D: Instance<Inst = DMA>,

Set the clock gate for the DMA controller

pub fn clock_gate_adc<A>(&self, adc: &A) -> ClockGate where
A: Instance<Inst = ADC>,

Returns the clock gate setting for the ADC

pub fn set_clock_gate_adc<A>(&mut self, adc: &mut A, gate: ClockGate) where
A: Instance<Inst = ADC>,

Set the clock gate for the ADC peripheral

pub fn clock_gate_pwm<P>(&self, pwm: &P) -> ClockGate where
P: Instance<Inst = PWM>,

Returns the clock gate setting for the ADC

pub fn set_clock_gate_pwm<P>(&mut self, pwm: &mut P, gate: ClockGate) where
P: Instance<Inst = PWM>,

Set the clock gate for the PWM peripheral

pub fn set_frequency_arm(&mut self, hz: u32) -> (ARMClock, IPGClock)

Set the ARM clock frequency, returning the new ARM and IPG clock frequency Changing this at runtime will affect anything that's using the ARM or IPG clocks as inputs. Keep this in mind when changing the core clock frequency throughout your programs.

pub fn frequency_arm(&self) -> (ARMClock, IPGClock)

Returns the ARM and IPG clock frequencies

impl<C> CCM<C> where
C: Clocks,
C::PIT: Instance<Inst = PIT>,
C::GPT: Instance<Inst = GPT>,

pub fn perclock(&self) -> &PerClock<C::PIT, C::GPT>

Returns a reference to the periodic clock

pub fn perclock_mut(&mut self) -> &mut PerClock<C::PIT, C::GPT>

Returns a mutable reference to the periodic clock

impl<C> CCM<C> where
C: Clocks,
C::I2C: Instance<Inst = I2C>,

pub fn i2c_clock(&self) -> &I2CClock<C::I2C>

Returns a reference to the I2C clock

pub fn i2c_clock_mut(&mut self) -> &mut I2CClock<C::I2C>

Returns a mutable reference to the I2C clock

impl<C> CCM<C> where
C: Clocks,
C::SPI: Instance<Inst = SPI>,

pub fn spi_clock(&self) -> &SPIClock<C::SPI>

Returns a reference to the SPI clock

pub fn spi_clock_mut(&mut self) -> &mut SPIClock<C::SPI>

Returns a mutable reference to the SPI clock

impl<C> CCM<C> where
C: Clocks,
C::UART: Instance<Inst = UART>,

pub fn uart_clock(&self) -> &UARTClock<C::UART>

Returns a reference to the UART clock

pub fn uart_clock_mut(&mut self) -> &mut UARTClock<C::UART>

Returns a mutable reference to the uart clock