Trait imxrt_ccm::Instance

pub unsafe trait Instance {
    type Inst: ClockGateLocator;
    pub fn instance(&self) -> Self::Inst;

    pub fn is_valid(inst: Self::Inst) -> bool;
}

A peripheral instance that has a clock gate

Instance lets you associate a peripheral with its clock gate. This lets you control a peripheral's clock gate by supplying the peripheral itself, rather than modifying an arbitrary field in a CCM register.

Implementers must hold the invariant: the return of instance, passed into is_valid, must be true. If instance never returns a variant, is_valid must return false.

Suppose that an i.MX RT processor only supports one GPT instance. Here's an example of a valid Instance implementation. It's assumed that there is only one MyGPT object, and that it represents the GPT1 timer.

struct MyGPT;
unsafe impl Instance for MyGPT {
    type Inst = GPT;
    fn instance(&self) -> Self::Inst {
        GPT::GPT1
    }
    fn is_valid(inst: GPT) -> bool {
        inst == GPT::GPT1
    }
}

let my_gpt = MyGPT;
assert!(MyGPT::is_valid(my_gpt.instance()));

If is_valid returned false when called with GPT::GPT1, the implementation is invalid.

unsafe impl Instance for MyGPT {
    type Inst = GPT;
    fn instance(&self) -> Self::Inst {
        GPT::GPT1
    }
    fn is_valid(inst: GPT) -> bool {
        false // Invalid implementation!
    }
}

Safety

You should only implement Instance on a true i.MX RT peripheral instance. Instances are only used when you have both a mutable reference to the instance, and a mutable reference to the CCM Handle. An incorrect implementation will let you control global, mutable state that should not be associated with the object.

Associated Types

type Inst: ClockGateLocator

An identifier that describes the instance

Loading content...
Required methods

pub fn instance(&self) -> Self::Inst

Returns the peripheral instance identifier

pub fn is_valid(inst: Self::Inst) -> bool

Returns true if this instance is valid for a particular implementation

Loading content...